% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection}
\alias{get_collection}
\title{Get all entries in a collection}
\usage{
get_collection(url, collection, limit = 10, token, rate = 50/60)
}
\arguments{
\item{url}{Website URL}

\item{collection}{Collection name}

\item{limit}{Number of item by page}

\item{token}{Bearer token}

\item{rate}{Rate limit, default to 50 per minute}
}
\value{
A dataframe with users data
}
\description{
Get all entries in a collection
}
