% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{get_orders}
\alias{get_orders}
\title{Get all orders}
\usage{
get_orders(url, limit = 100, token, rate = 50/60)
}
\arguments{
\item{url}{Website URL}

\item{limit}{Number of item by page}

\item{token}{Bearer token}

\item{rate}{Rate limit, default to 50 per minute}
}
\value{
A dataframe with orders data
}
\description{
Get all orders
}
