% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{get_orders_page}
\alias{get_orders_page}
\title{Get orders for one page}
\usage{
get_orders_page(url, page, limit, token, rate = 50/60)
}
\arguments{
\item{url}{Website URL}

\item{page}{Page number}

\item{limit}{Number of item by page}

\item{token}{Bearer token}

\item{rate}{Rate limit, default to 50 per minute}
}
\description{
Get orders for one page
}
\keyword{internal}
