% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection_list}
\alias{get_collection_list}
\title{Get ID and titles of items in collection}
\usage{
get_collection_list(url, collection, custom_fields = "", token)
}
\arguments{
\item{url}{Website URL}

\item{collection}{Collection name}

\item{custom_fields}{Custom fiels added, separated by , (also for the start)}

\item{token}{Bearer token}
}
\value{
A dataframe with fields
}
\description{
Get ID and titles of items in collection
}
