% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection_page}
\alias{get_collection_page}
\title{Get collection for one page}
\usage{
get_collection_page(url, collection, page, limit)
}
\arguments{
\item{url}{Website URL}

\item{collection}{Collection name}

\item{page}{Page number}

\item{limit}{Number of item by page}
}
\description{
Get collection for one page
}
\keyword{internal}
