% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection_params}
\alias{get_collection_params}
\title{Get number of items in collection}
\usage{
get_collection_params(url, collection)
}
\arguments{
\item{url}{Website URL}

\item{collection}{Collection name}
}
\description{
Get number of items in collection
}
\keyword{internal}
